# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "API node task",
      "title": "Node Task"

    },
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/api/category/{id}": {
            "delete": {
                "tags": [
                  "category"
                ],
                "summary": "Delete a category",
                "description": "",
                "operationId": "deleteCategory",
                "produces": [
                  "application/xml",
                  "application/json"
                ],
                "parameters": [
                  {
                    "name": "id",
                    "in": "path",
                    "description": "category id to delete",
                    "required": true,
                    "type": "string"
                  }
                ],
                "responses": {
                  "400": {
                    "description": "Invalid ID supplied"
                  },
                  "404": {
                    "description": "Category not found"
                  }
                }
              }
      },
...
    },
...
  }